+++ convert/src/main/kotlin/com/github/zeldigas/text2confl/convert/markdown/MarkdownConfiguration.kt
@@ -1,0 +1,6 @@
+package com.github.zeldigas.text2confl.convert.markdown
+
+data class MarkdownConfiguration(
+    val parseAnyMacro:Boolean = true,
+    val supportedMacros:List<String> = emptyList()
+)